rename object store to kv store
