+++ fastly/fixtures/object_store/create-store.yaml
@@ -2,7 +2,7 @@
 version: 1
 interactions:
 - request:
-    body: '{"name":"object-store-test-store"}'
+    body: '{"name":"kv-store-test-store"}'
     form: {}
     headers:
       Accept:
@@ -11,10 +11,10 @@ interactions:
       - application/json
       User-Agent:
       - FastlyGo/6.5.0 (+github.com/fastly/go-fastly; go1.19)
-    url: https://api.fastly.com/resources/stores/object
+    url: https://api.fastly.com/resources/stores/kv
     method: POST
   response:
-    body: '{"id":"fff9gl14gebt6m9kkp54ve","name":"object-store-test-store","created_at":"2022-09-15T17:55:35.881Z","updated_at":"2022-09-15T17:55:35.881Z"}'
+    body: '{"id":"fff9gl14gebt6m9kkp54ve","name":"kv-store-test-store","created_at":"2022-09-15T17:55:35.881Z","updated_at":"2022-09-15T17:55:35.881Z"}'
     headers:
       Accept-Ranges:
       - bytes
