+++ fastly/fixtures/object_store/get-store.yaml
@@ -7,10 +7,10 @@ interactions:
     headers:
       User-Agent:
       - FastlyGo/6.5.0 (+github.com/fastly/go-fastly; go1.19)
-    url: https://api.fastly.com/resources/stores/object/fff9gl14gebt6m9kkp54ve
+    url: https://api.fastly.com/resources/stores/kv/fff9gl14gebt6m9kkp54ve
     method: GET
   response:
-    body: '{"id":"fff9gl14gebt6m9kkp54ve","name":"object-store-test-store","created_at":"2022-09-15T17:55:36.122Z","updated_at":"2022-09-15T17:55:36.122Z"}'
+    body: '{"id":"fff9gl14gebt6m9kkp54ve","name":"kv-store-test-store","created_at":"2022-09-15T17:55:36.122Z","updated_at":"2022-09-15T17:55:36.122Z"}'
     headers:
       Accept-Ranges:
       - bytes
