+++ fastly/fixtures/resources/create-object-store.yaml
@@ -2,7 +2,7 @@
 version: 1
 interactions:
 - request:
-    body: '{"name":"test-object-store"}'
+    body: '{"name":"test-kv-store"}'
     form: {}
     headers:
       Accept:
@@ -11,10 +11,10 @@ interactions:
       - application/json
       User-Agent:
       - FastlyGo/7.0.0 (+github.com/fastly/go-fastly; go1.16.15)
-    url: https://api.fastly.com/resources/stores/object
+    url: https://api.fastly.com/resources/stores/kv
     method: POST
   response:
-    body: '{"id":"2dem3bjbpie7k1vyy70fh5","name":"test-object-store","created_at":"2023-01-19T11:28:13.754Z","updated_at":"2023-01-19T11:28:13.754Z"}'
+    body: '{"id":"2dem3bjbpie7k1vyy70fh5","name":"test-kv-store","created_at":"2023-01-19T11:28:13.754Z","updated_at":"2023-01-19T11:28:13.754Z"}'
     headers:
       Accept-Ranges:
       - bytes
