+++ fastly/fixtures/resources/create.yaml
@@ -2,10 +2,10 @@
 version: 1
 interactions:
 - request:
-    body: name=test-object-store-name-for-linking&resource_id=2dem3bjbpie7k1vyy70fh5
+    body: name=test-kv-store-name-for-linking&resource_id=2dem3bjbpie7k1vyy70fh5
     form:
       name:
-      - test-object-store-name-for-linking
+      - test-kv-store-name-for-linking
       resource_id:
       - 2dem3bjbpie7k1vyy70fh5
     headers:
@@ -16,7 +16,7 @@ interactions:
     url: https://api.fastly.com/service/7i6HN3TK9wS159v2gPAZ8A/version/5/resource
     method: POST
   response:
-    body: '{"id":"rn6c7YSyRxBJMejpyvoQa1","name":"test-object-store-name-for-linking","service_id":"7i6HN3TK9wS159v2gPAZ8A","version":5,"created_at":"2023-01-19T11:28:14Z","updated_at":"2023-01-19T11:28:14Z","deleted_at":null,"resource_id":"2dem3bjbpie7k1vyy70fh5","resource_type":"object-store","href":null}'
+    body: '{"id":"rn6c7YSyRxBJMejpyvoQa1","name":"test-kv-store-name-for-linking","service_id":"7i6HN3TK9wS159v2gPAZ8A","version":5,"created_at":"2023-01-19T11:28:14Z","updated_at":"2023-01-19T11:28:14Z","deleted_at":null,"resource_id":"2dem3bjbpie7k1vyy70fh5","resource_type":"object-store","href":null}'
     headers:
       Accept-Ranges:
       - bytes
