+++ fastly/fixtures/resources/update.yaml
@@ -2,10 +2,10 @@
 version: 1
 interactions:
 - request:
-    body: name=new-object-store-alias-for-my-service
+    body: name=new-kv-store-alias-for-my-service
     form:
       name:
-      - new-object-store-alias-for-my-service
+      - new-kv-store-alias-for-my-service
     headers:
       Content-Type:
       - application/x-www-form-urlencoded
@@ -14,7 +14,7 @@ interactions:
     url: https://api.fastly.com/service/7i6HN3TK9wS159v2gPAZ8A/version/5/resource/rn6c7YSyRxBJMejpyvoQa1
     method: PUT
   response:
-    body: '{"id":"rn6c7YSyRxBJMejpyvoQa1","name":"new-object-store-alias-for-my-service","service_id":"7i6HN3TK9wS159v2gPAZ8A","version":5,"created_at":"2023-01-19T11:28:14Z","updated_at":"2023-01-19T11:28:14Z","deleted_at":null,"resource_id":"2dem3bjbpie7k1vyy70fh5","resource_type":"object-store","href":null}'
+    body: '{"id":"rn6c7YSyRxBJMejpyvoQa1","name":"new-kv-store-alias-for-my-service","service_id":"7i6HN3TK9wS159v2gPAZ8A","version":5,"created_at":"2023-01-19T11:28:14Z","updated_at":"2023-01-19T11:28:14Z","deleted_at":null,"resource_id":"2dem3bjbpie7k1vyy70fh5","resource_type":"object-store","href":null}'
     headers:
       Accept-Ranges:
       - bytes
